Added Password Reset form to Login Popup
- Moved Login form into its own component
- Moved password reset into its own form component
- Added password reset form to Login Popup
