+++ frontend/.env
@@ -1,1 +1,1 @@
-REACT_APP_URL=http://remindus.web.app
+REACT_APP_URL=http://localhost:3000
